app/requirements.txt
fastapi
uvicorn[standard]
openai>=1.40.0
pydantic>=2.6
python-dotenv
bcrypt>=5.0.0
